fix: remove styles from separte files in pages
