--- src/pages/reset/styles.ts
@@ -1,74 +1,0 @@
-import { makeStyles } from '@material-ui/core'
-
-const useStyles = makeStyles(theme => ({
-  root: {
-    display: 'flex',
-    alignItems: 'center',
-    justifyContent: 'center',
-    width: '100vw',
-    height: '100vh'
-  },
-  container: {
-    width: '50vw',
-    height: '70vh',
-    display: 'flex',
-    borderRadius: 16,
-    [theme.breakpoints.down('sm')]: {
-      width: '90vw',
-      height: '90vh'
-    },
-    boxShadow: '2px 5px 10px 1px #eee'
-  },
-  leftImage: {
-    width: '60%',
-    backgroundColor: theme.palette.background.paper,
-    padding: 20,
-    [theme.breakpoints.down('sm')]: {
-      display: 'none'
-    },
-    borderTopLeftRadius: 16,
-    borderBottomLeftRadius: 16
-  },
-  rightForm: {
-    width: '100%',
-    display: 'flex',
-    alignItems: 'center',
-    justifyContent: 'space-around',
-    flexDirection: 'column',
-    backgroundColor: theme.palette.background.paper,
-    borderTopRightRadius: 16,
-    borderBottomRightRadius: 16,
-    padding: '50px',
-    [theme.breakpoints.down('sm')]: {
-      padding: '20px'
-    }
-  },
-  formContent: {
-    display: 'flex',
-    flexDirection: 'column',
-    alignItems: 'flex-start',
-    width: '100%'
-  },
-  inputText: {
-    margin: '10px 0',
-    width: '100%'
-  },
-  buttonsActions: {
-    display: 'flex',
-    width: '100%',
-    justifyContent: 'space-between',
-    [theme.breakpoints.down('sm')]: {
-      flexDirection: 'column',
-      height: 100
-    }
-  },
-  optionsFooter: {
-    display: 'flex',
-    alignItems: 'center',
-    [theme.breakpoints.down('sm')]: {
-      flexDirection: 'column'
-    }
-  }
-}))
-
-export default useStyles
